Create README.md
Kivy is a multiplatform GUI library, known for being responsive. It provides management of multiple screens in a single application. 

In this application we will be using multiple screens to log in user’s info and validate it. We will save the information in a csv file and use pandas to validate the information inside of the csv file by reading it into a DataFrame. To build the GUI we will use .kv file.  

Approach :

There will be three screens, one for letting the user log in, second for signing up and third for telling if the login was successful.
The info will be stored in a csv file.
Pandas Library will be used to read the csv file into a DataFrame and further check if the user info already exists or not.
If the information entered is invalid, popups will inform the user.
Finally, the user will be informed if the login was successful or not.